refactor(api): rename: getMockKeyData.ts → keyData.ts
